% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My R package_Multivariate Analysis.R
\name{max_cancor}
\alias{max_cancor}
\title{Maximum Canonical Correlation}
\usage{
max_cancor(X, Y)
}
\arguments{
\item{X}{matrix or dataframe, the first set of variables.}

\item{Y}{matrix or dataframe, the second set of variables.}
}
\value{
numeric, the maximum canonical correlation.
}
\description{
This function calculates the maximum canonical correlation between two sets of variables.
}
\examples{
# Create example datasets
set.seed(1)
X <- matrix(rnorm(100*5), 100, 5)
Y <- matrix(rnorm(100*3), 100, 3)

# Calculate the maximum canonical correlation
max_CC <- max_cancor(X, Y)
print(max_CC)
}
