% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My R package_Multivariate Analysis.R
\name{null_dist_CCA}
\alias{null_dist_CCA}
\title{Null Distribution for Canonical Correlation Analysis}
\usage{
null_dist_CCA(X, Y, num_permutations = 1000)
}
\arguments{
\item{X}{matrix or dataframe, a set of explanatory variables.}

\item{Y}{matrix or dataframe, a set of response variables.}

\item{num_permutations}{numeric, the number of permutations to perform. Default is 1000.}
}
\value{
density object representing the null distribution of permuted canonical correlations.
}
\description{
This function generates the null distribution of the maximum canonical correlation
by permuting the rows of the response variables and performing Canonical Correlation Analysis (CCA).
}
\examples{
# Example with random data
set.seed(1)
X <- matrix(rnorm(100*3), 100, 3)
Y <- matrix(rnorm(100*2), 100, 2)
null_dist_CCA(X, Y, num_permutations = 500)
}
