% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My R package_Multivariate Analysis.R
\name{Perm_test_CCA}
\alias{Perm_test_CCA}
\title{Permutation Test for Canonical Correlation Analysis}
\usage{
Perm_test_CCA(X, Y, num_permutations = 1000, observed)
}
\arguments{
\item{X}{matrix or dataframe, a set of explanatory variables.}

\item{Y}{matrix or dataframe, a set of response variables.}

\item{num_permutations}{numeric, the number of permutations to perform. Default is 1000.}

\item{observed}{numeric, the observed canonical correlation.}
}
\value{
numeric, the p-value representing the significance of the observed canonical correlation.
}
\description{
This function performs a permutation test to calculate the p-value of the observed canonical correlation
by comparing it to a null distribution generated by permuting the response variables.
}
\examples{
# Example with random data
set.seed(1)
X <- matrix(rnorm(100*3), 100, 3)
Y <- matrix(rnorm(100*2), 100, 2)
observed <- max(stats::cancor(X, Y)$cor)
p_value <- Perm_test_CCA(X, Y, num_permutations = 500, observed = observed)
}
