% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My R package_Multivariate Analysis.R
\name{simulate_multivariate_data}
\alias{simulate_multivariate_data}
\title{Simulate Multivariate Data}
\usage{
simulate_multivariate_data(n, means, diag_cor, within_cor, between_cor, column)
}
\arguments{
\item{n}{numeric, a single number specifying the sample size.}

\item{means}{numeric, a vector specifying the means for each variable.}

\item{diag_cor}{numeric, a single number specifying the correlation of each
variable with itself}

\item{within_cor}{numeric, a single number specifying the correlations between
variables within each multivariate outcome}

\item{between_cor}{numeric. a single number specifying the correlations between
variables from MV1 and variables from MV2}

\item{column, }{a single number specifying the column at which to split the
overall data into two separate multivariate datasets.}
}
\value{
A list containing both multivariate outcomes as separate elements
}
\description{
This function creates multivariate data with specified means and correlations.
}
\examples{
# Create a list with 2 multivariate outcomes. Each has 2 variables and 100 subjects
n <- 100
means <- rep(0, 4)
diag_cor  <- 1
within_cor <- 0.1
between_cor <- 0.2
column <- 2
my_data <- simulate_multivariate_data(n, means, diag_cor, within_cor, between_cor, column)
#  Access the split data
MV1 <- my_data$MV1
MV2 <- my_data$MV2
}
